html/rendering/replaced-elements/attributes-for-embedded-content-and-images/object-aspect-ratio.tentative.html

<!doctype html>
<title>Object width and height attributes are used to infer aspect-ratio</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  object {
    width: 100%;
    max-width: 100px;
    height: auto;
  }
</style>
<object data="/images/green.png"></object>
<object data="/images/green.png" width=100 height=125></object>
<script>
let t = async_test("Object width and height attributes are used to infer aspect-ratio");
function assert_ratio(obj, expected) {
  let epsilon = 0.001;
  assert_approx_equals(parseInt(getComputedStyle(obj).width, 10) / parseInt(getComputedStyle(obj).height, 10), expected, epsilon);
}
// Create and append a new object and immediately check the ratio.
// This is not racy because the spec requires the user agent to queue a task.
t.step(function() {
  var obj = document.createElement("object");
  obj.width = 250;
  obj.height = 100;
  obj.data = "/images/blue.png";
  document.body.appendChild(obj);
  assert_ratio(obj, 2.5);

  obj = document.createElement("object");
  obj.setAttribute("width", "0.8");
  obj.setAttribute("height", "0.2");
  obj.data = "/images/blue.png";
  document.body.appendChild(obj);
  assert_ratio(obj, 4);

  obj = document.createElement("object");
  obj.setAttribute("width", "50%");
  obj.setAttribute("height", "25%");
  obj.data = "/images/blue.png";
  document.body.appendChild(obj);
  // Percentages should be ignored.
  assert_equals(getComputedStyle(obj).height, "0px");
});

onload = t.step_func_done(function() {
  let objects = document.querySelectorAll("object");
  assert_ratio(objects[2], 1.266); // 1.266 is the original aspect ratio of blue.png
  assert_ratio(objects[1], 2.0); // 2.0 is the original aspect ratio of green.png
  assert_ratio(objects[0], 2.0); // Loaded object's aspect ratio, at least by default, overrides width / height ratio.
});
</script>
